Simple Singleton Pattern implementation
The task in this project is after each time I restart the scene, to make the cube:
to rotate at different angles.
different speed,
various colors
random scale and
in different places on the screen.
I found it boring clicking on the buttons to  restart the scene in order to see the task completed. I decided to implement the Singleton Pattern to restart the game automatically.